WIP: Restructure settings handling
