+++ src/pyskindose/plotting/create_geometry_plot.py
@@ -4,7 +4,7 @@ from pyskindose import constants as c
 from pyskindose import position_patient_phantom_on_table
 from pyskindose.phantom_class import Phantom
 from pyskindose.plotting.plot_geometry import plot_geometry
-from pyskindose.settings_pyskindose import PyskindoseSettings
+from pyskindose.settings import PyskindoseSettings
 
 
 def create_geometry_plot(
