+++ src/pyskindose/settings/__init__.py
@@ -1,0 +1,1 @@
+from .settings_pyskindose import PyskindoseSettings, initialize_settings
