+++ src/pyskindose/settings/normalization_settings.py
@@ -1,0 +1,74 @@
+from typing import Dict, Any, Optional, List
+
+import pandas as pd
+
+from pyskindose.constants import (
+    KEY_NORMALIZATION_DETECTOR_SIDE_LENGTH,
+    KEY_NORMALIZATION_FIELD_SIZE_MODE,
+    KEY_NORMALIZATION_MANUFACTURER,
+    KEY_NORMALIZATION_MODELS,
+    KEY_RDSR_MANUFACTURER,
+    KEY_RDSR_MANUFACTURER_MODEL_NAME,
+)
+from pyskindose.helpers.create_attributes_string import create_attributes_string
+from pyskindose.settings.rotation_direction import RotationDirection
+from pyskindose.settings.translation_direction import TranslationDirection
+from pyskindose.settings.translation_offset import TranslationOffset
+
+
+class NormalizationSettings:
+    """A class to normalize RDSR for PySkinDose compliance.
+
+    Attributes
+    ----------
+    trans_offset : pyskindose.settings.translation_offset.TranslationOffset
+        See class variables of _TranslationOffset
+    trans_dir : pyskindose.settings.translation_direction.TranslationDirection
+        See class variables of _TranslationDirection
+    rot_dir : pyskindose.settings.rotation_direction.RotationDirection
+        See class variables of _RotationDirection
+    field_size_mode : str
+        method for calculating field size at image receptor plane.
+        Choose either "CFA" (collimated field area) or "ACD" (actual shutter
+        distance). For more info, see calculate_field_size in geom_calc.py.
+    detector_side_length : str
+        side length of active image receptor area in cm.
+
+    """
+    def __init__(self, normalization_settings: List[Dict[str, Any]]):
+        """Initialize class attributes."""
+        self.normalization_settings_list: List[Dict[str, Any]] = normalization_settings
+        self.trans_offset: TranslationOffset = TranslationOffset()
+        self.trans_dir: TranslationDirection = TranslationDirection()
+        self.rot_dir: RotationDirection = RotationDirection()
+        self.field_size_mode: Optional[str] = None
+        self.detector_side_length: Optional[str] = None
+
+        self.attrs_str = create_attributes_string(attrs_parent=self, object_name="plot", indent_level=0)
+
+    def update_used_settings(self, data_parsed: pd.DataFrame):
+        manufacturer = data_parsed[KEY_RDSR_MANUFACTURER][0].casefold()
+        model = data_parsed[KEY_RDSR_MANUFACTURER_MODEL_NAME][0].casefold()
+
+        setting = [
+            setting
+            for setting in self.normalization_settings_list
+            if manufacturer == setting[KEY_NORMALIZATION_MANUFACTURER].casefold()
+               and model in setting[KEY_NORMALIZATION_MODELS].casefold()
+        ]
+
+        if not setting:
+            raise NotImplementedError(f"Could not find settings for the given manufacturer and model ({manufacturer=}, {model=})")
+
+        setting = setting[0]
+
+        self.trans_offset.update_translation_offset(offset=setting.get("translation_offset"))
+        self.trans_dir.update_translation_direction(directions=setting.get("translation_direction"))
+        self.rot_dir.update_rotation_direction(directions=setting.get("rotation_direction"))
+        self.field_size_mode = setting[KEY_NORMALIZATION_FIELD_SIZE_MODE]
+        self.detector_side_length = setting[KEY_NORMALIZATION_DETECTOR_SIDE_LENGTH]
+
+        self.attrs_str = create_attributes_string(attrs_parent=self, object_name="normalization", indent_level=0)
+
+    def update_attrs_str(self):
+        self.attrs_str = create_attributes_string(attrs_parent=self, object_name="normalization", indent_level=0)
