+++ tests/manual_tests/manual_test_plot_procedure_beam_collimations.py
@@ -2,7 +2,7 @@ from base_dev_settings import DEVELOPMENT_PARAMETERS
 
 from pyskindose import constants
 from pyskindose.main import main
-from pyskindose.settings_pyskindose import PyskindoseSettings
+from pyskindose.settings import PyskindoseSettings
 
 settings = PyskindoseSettings(settings=DEVELOPMENT_PARAMETERS)
 settings.mode = constants.MODE_PLOT_PROCEDURE
