+++ tests/unittests/test_settings.py
@@ -5,7 +5,7 @@ P = Path(__file__).parent.parent
 sys.path.insert(1, str(P.absolute()))
 
 from manual_tests.base_dev_settings import DEVELOPMENT_PARAMETERS
-from pyskindose.settings_pyskindose import PyskindoseSettings
+from pyskindose.settings import PyskindoseSettings
 
 settings = PyskindoseSettings(settings=DEVELOPMENT_PARAMETERS)
